refactor: Split delete
